requirements.txt
fastapi==0.95.0
python-telegram-bot==20.0
uvicorn==0.22.0
requests==2.25.1
